+++ src/Actions/ManagesSites.php
@@ -75,6 +75,22 @@ trait ManagesSites
     }
 
     /**
+     * Add Site Aliases.
+     *
+     * @param  int  $serverId
+     * @param  int  $siteId
+     * @param  array  $aliases
+     * @return \Laravel\Forge\Resources\Site
+     */
+    public function addSiteAliases($serverId, $siteId, array $aliases)
+    {
+        return new Site(
+            $this->put("servers/$serverId/sites/$siteId/aliases", compact('aliases'))['site']
+            + ['server_id' => $serverId], $this
+        );
+    }
+
+    /**
      * Refresh the site token.
      *
      * @param  int  $serverId
@@ -269,7 +285,7 @@ trait ManagesSites
             return $this->retry($this->getTimeout(), function () use ($serverId, $siteId) {
                 $site = $this->site($serverId, $siteId);
 
-                return ! is_null($site->deploymentStatus) ? $site : null;
+                return is_null($site->deploymentStatus) ? $site : null;
             });
         }