+++ CHANGELOG.md
@@ -1,6 +1,36 @@
 # Release Notes
 
-## [Unreleased](https://github.com/laravel/forge-sdk/compare/v3.0.0...master)
+## [Unreleased](https://github.com/laravel/forge-sdk/compare/v3.3.1...3.x)
+
+
+## [v3.3.1 (2020-12-14)](https://github.com/laravel/forge-sdk/compare/v3.3.0...v3.3.1)
+
+### Fixed
+
+- Fix deployments `$wait` ([9d24051](https://github.com/laravel/forge-sdk/commit/9d24051ae1cf5fd28109713b7d7712fcd80e194b))
+
+
+## [v3.3.0 (2020-11-03)](https://github.com/laravel/forge-sdk/compare/v3.2.0...v3.3.0)
+
+### Added
+- PHP 8 Support ([#98](https://github.com/laravel/forge-sdk/pull/98))
+
+
+## [v3.2.0 (2020-09-22)](https://github.com/laravel/forge-sdk/compare/v3.1.0...v3.2.0)
+
+### Added
+- Add site aliases function ([#95](https://github.com/laravel/forge-sdk/pull/95))
+- Forge defaults as class constants ([#94](https://github.com/laravel/forge-sdk/pull/94))
+
+
+## [v3.1.0 (2020-09-01)](https://github.com/laravel/forge-sdk/compare/v3.0.0...v3.1.0)
+
+### Added
+- Add restart worker ([d62ecb4](https://github.com/laravel/forge-sdk/commit/d62ecb4b654b0fa5db1dc5e8cb0131bb1ef92d27))
+
+### Fixed
+- Fix delete action for RedirectRule ([#91](https://github.com/laravel/forge-sdk/pull/91))
+- Fix Worker resource delete method ([#90](https://github.com/laravel/forge-sdk/pull/90))
 
 
 ## [v3.0.0 (2020-08-25)](https://github.com/laravel/forge-sdk/compare/v2.2...v3.0.0)